- Support fixes for PHP 8
git-svn-id: https://plugins.svn.wordpress.org/duplicator/trunk@2439315 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ installer/dup-installer/ctrls/ctrl.s2.dbinstall.php
@@ -390,7 +390,7 @@ class DUPX_DBInstall
             while ($row = mysqli_fetch_row($result)) {
                 $found_tables[] = $row[0];
             }
-            if (count($found_tables) > 0) {
+            if ($found_tables != null && count($found_tables) > 0) {
                 mysqli_query($this->dbh, "SET FOREIGN_KEY_CHECKS = 0;");
                 foreach ($found_tables as $table_name) {
                     $sql    = "DROP TABLE `".mysqli_real_escape_string($this->dbh, $this->post['dbname'])."`.`".mysqli_real_escape_string($this->dbh, $table_name)."`";
